+++ poc/classes/AuxSTAxiom.py
@@ -18,6 +18,7 @@ class AuxSTAxiom(AuxSTBlock):
         self.keyword = ""
 
     def evaluate(self, sem):
+        sem.analyzer.current_building_block = self
         if self.constant_value() is None:
             signature = None
             for child in self.children:
@@ -54,3 +55,4 @@ class AuxSTAxiom(AuxSTBlock):
             # replace the stack by the immutable value
             sem.eval_stack.pop()
             sem.eval_stack.append(self.constant_value())
+        self.set_sc_ready()