+++ poc/classes/AuxSTClassInstance.py
@@ -22,4 +22,6 @@ class AuxSTClassInstance(AuxSTInstance):
         return new_class_instance
 
     def evaluate(self, sem):
+        sem.analyzer.current_building_block = self
         sem.eval_stack[-1].value = InbuiltUndefined()
+        self.set_sc_ready()