+++ poc/classes/AuxSTDefinitionPredicate.py
@@ -51,6 +51,7 @@ class AuxSTDefinitionPredicate(AuxSTBlock):
         self.filter_misused_templates(error_mgr, filename)
 
     def evaluate(self, sem):
+        sem.analyzer.current_building_block = self
         if self.constant_value() is None:
             signature = None
             for child in self.children:
@@ -77,3 +78,4 @@ class AuxSTDefinitionPredicate(AuxSTBlock):
             # replace the stack by the immutable value
             sem.eval_stack.pop()
             sem.eval_stack.append(self.constant_value())
+        self.set_sc_ready()