+++ poc/classes/AuxSTTheoremLikeStatementOrConjecture.py
@@ -14,6 +14,7 @@ class AuxSTTheoremLikeStatementOrConjecture(AuxSTBlock):
         self.zto = zto
 
     def evaluate(self, sem):
+        sem.analyzer.current_building_block = self
         if self.constant_value() is None:
             signature = None
             for child in self.children:
@@ -41,3 +42,4 @@ class AuxSTTheoremLikeStatementOrConjecture(AuxSTBlock):
             # replace the stack by the immutable value
             sem.eval_stack.pop()
             sem.eval_stack.append(self.constant_value())
+        self.set_sc_ready()