+++ poc/CHANGES.md
@@ -1,9 +1,16 @@
 # Changes in the FPL Parser and Interpreter
+## 1.8.6
+* Feature:
+  * New class AuxSelfContainment
+  * New class AuxParamsArgsMatcher
+  * Initial type matching algorithm for arguments matching parameters
+* Bugfixes:
+  * Some missing FplWrongArguments recognitions fixed
 ## 1.8.5
 * Refactoring:
   * Removing get_required_signature() and replacing string-based type recognition in evaluation methods by a AuxSTType-based recognition
 * Feature:
-  * New error FplVariableBound and som first related test cases
+  * New error FplVariableBound and some first related test cases
   * Initial implementation of the evaluate method of different symbol table node subtypes 
   * Default constructors inherit the position as the class declaration (affected expected test cases of symbol table adjusted accordingly)
   * Initial implementation of FplWrongArguments recognition